+++ src/groups/mqb/mqbblp/mqbblp_cluster.h
@@ -467,17 +467,17 @@ class Cluster : public mqbi::Cluster,
                                            const bmqp::QueueId& queueId,
                                            mqbc::ClusterNodeSession* ns,
                                            bmqp::EventType::Enum eventType);
-    // Validate a message of the specified 'eventType' using the specified
-    // 'queueId' and 'ns'. Return one of `ValidationResult` values. Populate
-    // the specified 'queueHandle' if the queue is found.
+    // Validate a message of the specified `eventType` using the specified
+    // `queueId` and `ns`. Return one of `ValidationResult` values. Populate
+    // the specified `queueHandle` if the queue is found.
 
     bool validateRelayMessage(mqbc::ClusterNodeSession** ns,
                               bsl::ostream*              errorStream,
                               const int                  pid);
-    // Validate a relay message using the specified 'pid'. Return true if the
-    // message is valid and false otherwise. Populate the specified 'ns' if the
+    // Validate a relay message using the specified `pid`. Return true if the
+    // message is valid and false otherwise. Populate the specified `ns` if the
     // message is valid or load a descriptive error message into the
-    // 'errorStream' if the message is invalid.
+    // `errorStream` if the message is invalid.
 
     /// Executes in any thread.
     void
@@ -539,10 +539,10 @@ class Cluster : public mqbi::Cluster,
     // CREATORS
 
     /// Create a new object representing a cluster having the specified
-    /// 'name', 'clusterConfig' and 'statContexts', associated to the
-    /// specified 'netCluster' and using the specified 'domainFactory',
-    /// 'scheduler', 'dispatcher', 'transportManager', and 'resources'.  Use
-    /// the specified 'allocator' for any memory allocation.
+    /// `name`, `clusterConfig` and `statContexts`, associated to the
+    /// specified `netCluster` and using the specified `domainFactory`,
+    /// `scheduler`, `dispatcher`, `transportManager`, and `resources`.  Use
+    /// the specified `allocator` for any memory allocation.
     Cluster(const bslstl::StringRef&                      name,
             const mqbcfg::ClusterDefinition&              clusterConfig,
             bslma::ManagedPtr<mqbnet::Cluster>            netCluster,
