+++ src/groups/mqb/mqbblp/mqbblp_queueconsumptionmonitor.t.cpp
@@ -108,7 +108,6 @@ struct Test : mwctst::Test {
     mqbmock::Cluster                          d_cluster;
     mqbmock::Domain                           d_domain;
     mqbmock::Queue                            d_queue;
-    mqbi::ClusterResources                    d_resources;
     QueueState                                d_queueState;
     QueueConsumptionMonitor                   d_monitor;
     mqbs::InMemoryStorage                     d_storage;
@@ -146,14 +145,13 @@ Test::Test()
 , d_cluster(&d_bufferFactory, s_allocator_p)
 , d_domain(&d_cluster, s_allocator_p)
 , d_queue(&d_domain, s_allocator_p)
-, d_resources()
 , d_queueState(&d_queue,
                d_uri,
                d_id,
                d_storageKey,
                d_partitionId,
                &d_domain,
-               d_resources,
+               d_cluster._resources(),
                s_allocator_p)
 , d_monitor(&d_queueState, s_allocator_p)
 , d_storage(d_queue.uri(),
