+++ src/groups/mqb/mqbblp/mqbblp_queueengineutil.h
@@ -407,29 +407,30 @@ struct QueueEngineUtil_AppState {
     /// Reset the internal state to have no consumers.
     void undoRouting();
 
-    /// Attempt to deliver all pending data up to the specified 'end' (the
+    /// Attempt to deliver all pending data up to the specified `end` (the
     /// iterator).  First, attempt to drain the Redelivery and PutAside Lists.
     /// While doing so, load the message delay into the specified `delay` and
     /// throttle the redelivery as in the case of a Poisonous message.
     /// If the Redelivery List is empty, attempt to deliver data starting from
-    /// 'start' (the resumePoint or the beginning of the stream) to the 'end'.
+    /// `start` (the resumePoint or the beginning of the stream) to the `end`.
     /// This is what any QueueEngine calls whenever there is a chance for the
     /// App to make progress: any configuration change, capacity availability.
     /// Note that depending upon queue's mode, messages are delivered either
     /// to all consumers (broadcast mode), or in a round-robin manner (every
     /// other mode).
+    /// Use the specified `reader` to read data for delivery.
     size_t deliverMessages(bsls::TimeInterval*          delay,
                            mqbi::StorageIterator*       reader,
                            mqbi::StorageIterator*       start,
                            const mqbi::StorageIterator* end);
 
-    /// Try to deliver to the next available consumer the specified 'message'.
+    /// Try to deliver to the next available consumer the specified `message`.
     /// If poisonous message handling requires a delay in the delivery, iterate
     /// all highest priority consumers, load the lowest delay into the
-    /// specified 'delay' and return 'e_DELAY.  If no delay is required, try to
-    /// send the 'message' to a highest priority consumer with matching
-    /// subscription.  Return corresponding result: 'e_SUCCESS',
-    /// 'e_NO_SUBSCRIPTION', 'e_NO_CAPACITY'. or 'e_NO_CAPACITY_ALL'.
+    /// specified `delay` and return `e_DELAY`.  If no delay is required, try
+    /// to send the `message` to a highest priority consumer with matching
+    /// subscription.  Return corresponding result: `e_SUCCESS`,
+    /// `e_NO_SUBSCRIPTION`, `e_NO_CAPACITY`. or `e_NO_CAPACITY_ALL`.
     Routers::Result tryDeliverOneMessage(bsls::TimeInterval*          delay,
                                          const mqbi::StorageIterator* message,
                                          bool isOutOfOrder);
@@ -515,17 +516,17 @@ struct QueueEngineUtil_AppState {
     /// pending data.
     void unauthorize();
 
-    /// Save the specified 'guid' in the PutAside list of messages for which
+    /// Save the specified `guid` in the PutAside list of messages for which
     /// there is no matching subscription.  The delivery of those messages will
     // be attempted upon configuration change.
     void putAside(const bmqt::MessageGUID& guid);
 
-    /// Save the specified 'guid' in the Redelivery list of messages.  These
+    /// Save the specified `guid` in the Redelivery list of messages.  These
     /// messages get delivered as Out-Of-Order.
     void putForRedelivery(const bmqt::MessageGUID& guid);
 
     /// Save the current position in the data stream to resume the delivery (by
-    /// 'deliverMessages').
+    /// `deliverMessages`).
     void setResumePoint(const bmqt::MessageGUID& guid);
 
     /// Return a reference offering modifiable access to the current routing
@@ -538,11 +539,11 @@ struct QueueEngineUtil_AppState {
 
     // ACCESSORS
 
-    /// Return 'true' if this App is not behind: authorized, empty Redelivery
+    /// Return `true` if this App is not behind: authorized, empty Redelivery
     /// List and no resume point.
     bool isReadyForDelivery() const;
 
-    /// Return 'true' if this App does not have any never delivered data before
+    /// Return `true` if this App does not have any never delivered data before
     /// the queue iterator: empty PutAside List and no resume point.
     bool isAtEndOfStorage() const;
 
@@ -565,7 +566,7 @@ struct QueueEngineUtil_AppState {
     /// Return the Id.
     const bsl::string& appId() const;
 
-    /// Return 'true' if this App is authorized.
+    /// Return `true` if this App is authorized.
     bool isAuthorized() const;
 
     /// Return the current resume point (empty when none).
@@ -575,7 +576,7 @@ struct QueueEngineUtil_AppState {
     /// routing state controlling evaluation of subscriptions.
     const bsl::shared_ptr<Routers::AppContext>& routing() const;
 
-    /// Report queue stats upon delivery of the specified 'message'.
+    /// Report queue stats upon delivery of the specified `message`.
     void reportStats(const mqbi::StorageIterator* message) const;
 };
 
@@ -596,7 +597,7 @@ struct QueueEngineUtil_AppsDeliveryContext {
     mqbi::Queue*                      d_queue_p;
     bsl::optional<bsls::Types::Int64> d_timeDelta;
     // Avoid reading the attributes if not necessary.  Get timeDelta on demand.
-    // See comment in 'QueueEngineUtil_AppsDeliveryContext::processApp'.
+    // See comment in `QueueEngineUtil_AppsDeliveryContext::processApp`.
 
   public:
     QueueEngineUtil_AppsDeliveryContext(mqbi::Queue*           queue,
@@ -606,24 +607,24 @@ struct QueueEngineUtil_AppsDeliveryContext {
     /// Prepare the context to process next message.
     void reset();
 
-    /// Return 'true' if the specified 'app' is not a broadcast app and has an
+    /// Return `true` if the specified `app` is not a broadcast app and has an
     /// available handle to deliver the current message with the specified
-    /// 'ordinal'.
-    /// 'false' return value indicates that the 'app' is either a  broadcast,
+    /// `ordinal`.
+    /// `false` return value indicates that the `app` is either a  broadcast,
     /// or unauthorized, or does not have matching subscription, or does not
     /// have the capacity for any existing subscription or did not drain its
     /// redelivery list, or is already behind.  In any case, an authorized app
     /// sets its resume point, so the queue iterator can continue to advance
-    /// unless in the 'e_NO_CAPACITY_ALL' case when all apps have no capacity
+    /// unless in the `e_NO_CAPACITY_ALL` case when all apps have no capacity
     /// for any existing subscription.
-    /// The  'ordinal' controls how we read the state from the data stream.  In
-    /// the case of 'RootQueueEngine' (Primary), the data stream is the storage
-    /// and the 'ordinal' is the App ordinal in the stream.
-    /// In the case of 'RelayQueueEngine', the stream is 'PushStream' and the
-    /// 'ordinal' is the offset in the received PUSH message.
+    /// The  `ordinal` controls how we read the state from the data stream.  In
+    /// the case of `RootQueueEngine` (Primary), the data stream is the storage
+    /// and the `ordinal` is the App ordinal in the stream.
+    /// In the case of `RelayQueueEngine`, the stream is `PushStream` and the
+    /// `ordinal` is the offset in the received PUSH message.
     bool processApp(QueueEngineUtil_AppState& app, unsigned int ordina);
 
-    /// Collect and prepare data for the subsequent 'deliverMessage' call.
+    /// Collect and prepare data for the subsequent `deliverMessage` call.
     bool visit(const Routers::Subscription* subscription,
                const mqbi::AppMessage&      appView);
     bool visitBroadcast(const Routers::Subscription* subscription);
@@ -631,12 +632,12 @@ struct QueueEngineUtil_AppsDeliveryContext {
     /// Deliver message to the previously processed handles.
     void deliverMessage();
 
-    /// Return 'true' if the delivery can continue iterating dataStream
-    /// The 'false' return value indicates either the end of the dataStream or
-    /// the the 'e_NO_CAPACITY_ALL' case.
+    /// Return `true` if the delivery can continue iterating dataStream
+    /// The `false` return value indicates either the end of the dataStream or
+    /// the the `e_NO_CAPACITY_ALL` case.
     bool doRepeat() const;
 
-    /// Return 'true' if there is at least one delivery target selected.
+    /// Return `true` if there is at least one delivery target selected.
     bool isEmpty() const;
 
     bsls::Types::Int64 timeDelta();
