+++ src/groups/mqb/mqbblp/package/mqbblp.mem
@@ -9,6 +9,7 @@ mqbblp_domain
 mqbblp_localqueue
 mqbblp_messagegroupidhelper
 mqbblp_messagegroupidmanager
+mqbblp_pushstream
 mqbblp_queue
 mqbblp_queueconsumptionmonitor
 mqbblp_queueenginetester
