+++ src/groups/mqb/mqbmock/mqbmock_cluster.h
@@ -157,6 +157,9 @@ class Cluster : public mqbi::Cluster {
     bslma::Allocator* d_allocator_p;
     // Allocator to use
 
+    bdlbb::BlobBufferFactory* d_bufferFactory_p;
+    // Buffer factory to use
+
     BlobSpPool d_blobSpPool;
     // Blob sp pool to use
 
@@ -165,8 +168,6 @@ class Cluster : public mqbi::Cluster {
 
     bdlmt::EventScheduler d_scheduler;
     // Event scheduler to use
-
-    mqbi::ClusterResources d_resources;
 
     bdlmt::EventSchedulerTestTimeSource d_timeSource;
     // Time source for the event scheduler
@@ -218,6 +219,8 @@ class Cluster : public mqbi::Cluster {
     // Dispatcher client data
 
     EventProcessor d_processor;
+
+    mqbi::ClusterResources d_resources;
 
   private:
     // NOT IMPLEMENTED
@@ -525,6 +528,7 @@ class Cluster : public mqbi::Cluster {
     const bdlmt::EventSchedulerTestTimeSource& _timeSource() const;
     const TestChannelMap&                      _channels() const;
     const mqbc::ClusterData*                   _clusterData() const;
+    const mqbi::ClusterResources&              _resources() const;
 
     /// Return the value of the corresponding member of this object.
     const mqbc::ClusterState& _state() const;
@@ -566,7 +570,7 @@ inline void Cluster::_setEventProcessor(const EventProcessor& processor)
 
 inline bdlbb::BlobBufferFactory* Cluster::_bufferFactory()
 {
-    return d_resources.d_bufferFactory_p;
+    return d_bufferFactory_p;
 }
 
 inline bdlmt::EventScheduler& Cluster::_scheduler()
@@ -654,6 +658,11 @@ inline const mqbc::ClusterData* Cluster::_clusterData() const
 inline const mqbc::ClusterState& Cluster::_state() const
 {
     return d_state;
+}
+
+inline const mqbi::ClusterResources& Cluster::_resources() const
+{
+    return d_resources;
 }
 
 inline bsls::TimeInterval Cluster::getTime() const
